+++ test/specs/delete.spec.js
@@ -40,7 +40,7 @@ describe('del', () => {
 
 		const test = dare.del('groups', {id: 20_000});
 
-		await assert.rejects(test, (error) => {
+		await assert.rejects(test, error => {
 			assert(error instanceof DareError);
 			assert.strictEqual(error.code, DareError.NOT_FOUND);
 			return true;
@@ -204,53 +204,5 @@ describe('del', () => {
 		});
 
 		assert.strictEqual(test.affectedRows, 1);
-	});
-
-	describe('DB Engine specific tests', () => {
-		const DB_ENGINE = 'postgres:16.3';
-		let dareInst;
-
-		beforeEach(() => {
-			dareInst = dare.use({engine: DB_ENGINE});
-		});
-
-		it(`${DB_ENGINE} should delete with subquery conditions rather than JOINs`, async () => {
-			dareInst.options.models = {
-				tbl: {
-					schema: {
-						// Create a reference to tblB
-						ref_id: ['tblB.id'],
-					},
-				},
-			};
-
-			dareInst.execute = async ({sql, values}) => {
-				sqlEqual(
-					sql,
-					`DELETE FROM tbl
-					WHERE tbl.id = ?
-						AND tbl.ref_id IN (
-							SELECT id FROM (
-								SELECT a.id FROM tblB a WHERE a.id= ?
-							) AS a_tmp
-						)
-					`
-				);
-				assert.deepStrictEqual(values, [1, 1]);
-				return {success: true};
-			};
-
-			const test = await dareInst.del({
-				table: 'tbl',
-				filter: {
-					id: 1,
-					tblB: {
-						id: 1,
-					},
-				},
-			});
-
-			assert.deepStrictEqual(test, {success: true});
-		});
 	});
 });