+++ test/specs/post.spec.js
@@ -277,9 +277,12 @@ describe('post', () => {
 					body: {name: given},
 				});
 
-				await assert.rejects(call, (error) => {
+				await assert.rejects(call, error => {
 					assert.ok(error instanceof DareError);
-					assert.match(error.message, /Field 'name' does not accept objects as values/);
+					assert.match(
+						error.message,
+						/Field 'name' does not accept objects as values/
+					);
 					assert.strictEqual(error.code, DareError.INVALID_VALUE);
 					return true;
 				});
@@ -311,48 +314,6 @@ describe('post', () => {
 					table: 'test',
 					body: {startDate: testValue},
 				});
-			});
-		});
-	});
-
-	describe('DB Engine specific tests', () => {
-		const DB_ENGINE = 'postgres:16.3';
-		let dareInst;
-
-		beforeEach(() => {
-			dareInst = dare.use({engine: DB_ENGINE});
-		});
-
-		it(`${DB_ENGINE} should use ON CONFLICT ... UPDATE ...`, async () => {
-			dareInst.execute = async ({sql, values}) => {
-				sqlEqual(
-					sql,
-					'INSERT INTO test ("id", "name") VALUES (?, ?) ON CONFLICT (id) DO UPDATE SET "name"=EXCLUDED."name" RETURNING id'
-				);
-				assert.deepStrictEqual(values, [1, 'name']);
-				return {success: true};
-			};
-
-			return dareInst.post({
-				table: 'test',
-				body: {id: 1, name: 'name'},
-				duplicate_keys_update: ['name'],
-			});
-		});
-		it(`${DB_ENGINE} should use ON CONFLICT DO NOTHING`, async () => {
-			dareInst.execute = async ({sql, values}) => {
-				sqlEqual(
-					sql,
-					'INSERT INTO test ("id", "name") VALUES (?, ?) ON CONFLICT DO NOTHING RETURNING id'
-				);
-				assert.deepStrictEqual(values, [1, 'name']);
-				return {success: true};
-			};
-
-			return dareInst.post({
-				table: 'test',
-				body: {id: 1, name: 'name'},
-				duplicate_keys: 'ignore',
 			});
 		});
 	});